+++ src/components/video-player/video-player.tsx
@@ -1,4 +1,5 @@
-import { FC, useRef, useEffect } from 'react';
+import HLS from 'hls.js';
+import { FC, useRef, useEffect, memo } from 'react';
 import { Typography, IconButton, Button } from '@mui/material';
 import { IconChevronLeft } from '@tabler/icons-react';
 import {
@@ -6,18 +7,21 @@ import {
   MediaPlayerInstance,
   MediaProvider,
   useMediaState,
+  MediaProviderAdapter,
   isHLSProvider, Track
 } from '@vidstack/react';
+
 import { DefaultVideoLayout, defaultLayoutIcons } from '@vidstack/react/player/layouts/default';
-import Label from '../label';
-import { useResponsive } from '@src/hooks/use-responsive';
 // @ts-ignore
 import '@vidstack/react/player/styles/default/theme.css';
 // @ts-ignore
 import '@vidstack/react/player/styles/default/layouts/audio.css';
 // @ts-ignore
 import '@vidstack/react/player/styles/default/layouts/video.css';
+
 import useGetSubtitles from '@src/hooks/use-get-subtitles.ts';
+import { useResponsive } from '@src/hooks/use-responsive';
+import Label from '../label';
 
 export type VideoPlayerProps = {
   src: string;
@@ -35,7 +39,7 @@ export const VideoPlayer: FC<VideoPlayerProps> = ({ src, cid, titleMovie, onBack
 
   useEffect(() => {
     if (cid) getSubtitles(cid);
-  }, [cid, getSubtitles]);
+  }, [cid]);
 
   useEffect(() => {
     const handleKeyDown = (event: KeyboardEvent) => {
@@ -43,105 +47,95 @@ export const VideoPlayer: FC<VideoPlayerProps> = ({ src, cid, titleMovie, onBack
     };
 
     document?.addEventListener('keydown', handleKeyDown);
-
-    return () => {
-      document?.removeEventListener('keydown', handleKeyDown);
-    };
-  }, [onBack]);
-
-  useEffect(() => {
-    if (player.current) {
-      if (isHLSProvider(player.current.provider)) {
-        player.current.provider.config = {
-          debug: false,
-          autoStartLoad: true,
-          capLevelOnFPSDrop: false,
-          capLevelToPlayerSize: false,
-          initialLiveManifestSize: 1,
-
-          maxBufferLength: 60,
-          maxMaxBufferLength: 600,
-          backBufferLength: Infinity,
-          frontBufferFlushThreshold: Infinity,
-          maxBufferSize: 80 * 1000 * 1000,
-          maxBufferHole: 0.2,
-          startPosition: -1,
-          nudgeOffset: 0.2,
-          nudgeMaxRetry: 5,
-
-          liveSyncDurationCount: 3,
-          liveSyncOnStallIncrease: 1,
-          liveMaxLatencyDurationCount: Infinity,
-          liveDurationInfinity: false,
-          preferManagedMediaSource: false,
-          enableWorker: true,
-          enableSoftwareAES: false,
-          startFragPrefetch: true,
-
-          testBandwidth: true,
-          progressive: false,
-          lowLatencyMode: false,
-
-          fpsDroppedMonitoringPeriod: 5000,
-          fpsDroppedMonitoringThreshold: 0.2,
-
-          enableDateRangeMetadataCues: false,
-          enableMetadataCues: false,
-          enableID3MetadataCues: false,
-          enableWebVTT: true,
-          enableIMSC1: false, // TODO change when subtitles needed
-          enableCEA708Captions: false, // TODO change when subtitles needed
-
-          stretchShortVideoTrack: false,
-          maxAudioFramesDrift: 1,
-          forceKeyFrameOnDiscontinuity: true,
-
-          abrEwmaFastLive: 3.0,
-          abrEwmaSlowLive: 9.0,
-          abrEwmaFastVoD: 2.0,
-          abrEwmaSlowVoD: 6.0,
-          abrEwmaDefaultEstimate: 1_000_000,
-          abrEwmaDefaultEstimateMax: 10_000_000,
-          abrBandWidthFactor: 0.9,
-          abrBandWidthUpFactor: 0.75,
-          abrMaxWithRealBitrate: true,
-
-          maxStarvationDelay: 3,
-          maxLoadingDelay: 3,
-          minAutoBitrate: 0,
-          emeEnabled: false,
-          licenseXhrSetup: undefined,
-          drmSystems: {},
-          drmSystemOptions: {},
-          fragLoadPolicy: {
-            default: {
-              maxTimeToFirstByteMs: 5000,
-              maxLoadTimeMs: 60_000,
-              timeoutRetry: {
-                maxNumRetry: 3,
-                retryDelayMs: 1000,
-                maxRetryDelayMs: 5000,
-              },
-              errorRetry: {
-                maxNumRetry: 5,
-                retryDelayMs: 2000,
-                maxRetryDelayMs: 10000,
-                backoff: 'exponential',
-              },
-            },
-          },
-        };
-      }
-    }
+    return () => document?.removeEventListener('keydown', handleKeyDown);
   }, []);
+
+  // on provider (HLS) initialization
+  const onProviderSetup = (provider: MediaProviderAdapter) => {
+    if (isHLSProvider(provider)) {
+      provider.instance?.on(HLS.Events.ERROR, (_, data: any) => {
+        if (data.details === HLS.ErrorDetails.BUFFER_STALLED_ERROR) {
+          console.log("Seek Stalling Detected, Adjusting Buffer...");
+          provider.instance?.startLoad();
+        }
+
+        if (data.fatal && data.type === HLS.ErrorTypes.MEDIA_ERROR) {
+          console.warn("Recovering from Media Error...");
+          provider.instance?.recoverMediaError();
+        }
+      });
+    }
+  }
+
+  // when the provider has changed, setup config..
+  const onProviderChange = (provider: MediaProviderAdapter | null) => {
+    if (isHLSProvider(provider)) {
+      provider.library = HLS;
+      provider.config = {
+        // https://github.com/video-dev/hls.js/blob/master/docs/API.md
+        // maxBufferLength defines the target amount of video (in seconds) the player tries to keep buffered.
+        // The buffer plays a crucial role in balancing playback stability and adaptive bitrate (ABR) decisions.
+        // A larger buffer reduces rebuffering risk but may delay quality switches, while a smaller buffer
+        // allows faster adaptation but increases the chance of playback interruptions.
+        // Finding the right balance ensures smooth playback without unnecessary network congestion.
+        // (hls_time = 6 + maxBufferLength = 30) = 5 fragments in buffer
+        "maxBufferLength": 60, // Max video buffer length in seconds
+        "maxMaxBufferLength": 600, // Absolute max buffer length
+        // maxStarvationDelay defines the maximum acceptable time (in seconds) a fragment can take to download 
+        // while playback is already in progress.
+        // - If a fragment is estimated to take longer than this value and the buffer is running low, 
+        //   the player switches the best quality that matches this time constraint.
+        // - This ensures a continuous playback experience by adapting the quality to network conditions in real-time.
+        "maxStarvationDelay": 2,
+        // maxLoadingDelay defines the maximum allowed time (in seconds) to load the initial fragments when starting playback.
+        // - The ABR controller ensures:
+        //   - The time to fetch the first low-quality fragment (e.g., 420p)
+        //   - + The time to fetch the estimated optimal-quality fragment (e.g., 720p)
+        //   - is below this value.
+        // - If the total loading time exceeds maxLoadingDelay, the player starts with a lower quality 
+        //   to minimize startup delay and ensure fast playback.
+        // - Unlike maxStarvationDelay, this setting only applies at the **start** of playback,
+        //   ensuring the video loads quickly even if it means initially using a lower quality.
+        "maxLoadingDelay": 3,
+        // abrEwmaFastVod: Controls how quickly the algorithm reacts to bandwidth changes in VOD (Video On Demand).
+        // A higher value makes the algorithm less sensitive to short-term fluctuations, smoothing out rapid changes.
+        // Recommended range: 2.0 - 5.0 (Higher = Smoother)
+        "abrEwmaFastVoD": 2,
+        // abrEwmaSlowVod: Controls the long-term average bandwidth estimation for adaptive bitrate switching.
+        // A higher value averages the bandwidth over a longer period, reducing frequent quality switches.
+        // Recommended range: 10.0 - 20.0 (Higher = More stable, but slower adaptation)
+        "abrEwmaSlowVoD": 9,
+        // abrBandWidthFactor: Determines how conservatively HLS estimates available bandwidth.
+        // A value < 1.0 ensures HLS.js does not use the full estimated bandwidth, preventing aggressive quality changes.
+        // Recommended range: 0.7 - 0.9 (Lower = More cautious, fewer quality switches)
+        "abrBandWidthFactor": 0.9,
+        // abrBandWidthUpFactor: Controls how aggressively the player upgrades to a higher bitrate.
+        // A lower value prevents HLS.js from switching to a higher quality too quickly, reducing unnecessary upscaling.
+        // Recommended range: 0.5 - 0.8 (Lower = More stable, avoids excessive upscaling)
+        "abrBandWidthUpFactor": 0.7,
+        "enableSoftwareAES": false, // Disable software AES decryption
+        "enableID3MetadataCues": false, // Disable ID3 metadata cues
+        "enableWebVTT": true, // Enable WebVTT subtitles
+        "enableIMSC1": false, // Disable IMSC1 subtitles
+        "enableCEA708Captions": false, // Disable CEA-708 captions,
+        "lowLatencyMode": false,
+        "enableWorker": true,
+        "nudgeOffset": 0.4,
+        "nudgeMaxRetry": 2,
+        // "maxFragLookUpTolerance": 0.4,
+        "startFragPrefetch": true
+      };
+
+
+    }
+  }
 
   return (
     <MediaPlayer
       ref={player}
-      src={{
-        src,
-        type: 'application/x-mpegurl',
-      }}
+      src={{ src, type: 'application/x-mpegurl' }}
+      onProviderChange={onProviderChange}
+      onProviderSetup={onProviderSetup}
       viewType="video"
       streamType="on-demand"
       logLevel="warn"
@@ -202,4 +196,6 @@ export const VideoPlayer: FC<VideoPlayerProps> = ({ src, cid, titleMovie, onBack
   );
 };
 
-export default VideoPlayer;
+export default memo(VideoPlayer, (prevProps, nextProps) => {
+  return prevProps.cid === nextProps.cid;
+});