+++ vite.config.ts
@@ -1,29 +1,36 @@
+/// <reference types="vitest" />
 import path from 'path';
 import react from '@vitejs/plugin-react';
 import { defineConfig, loadEnv } from 'vite';
 import { sentryVitePlugin } from "@sentry/vite-plugin";
 import preserveDirectives from 'rollup-preserve-directives'
+import { codecovVitePlugin } from "@codecov/vite-plugin";
 import { nodePolyfills } from 'vite-plugin-node-polyfills'
 
 export default defineConfig(({ mode }) => {
   // Load environment variables based on the current mode
   const env = loadEnv(mode, process.cwd(), '');
   const pure = mode === 'production' ? ['console.log', 'console.info', 'console.warn'] : []
-  
+
   return {
     esbuild: { pure },
     plugins: [
       react(),
       preserveDirectives(),
       sentryVitePlugin({
-        authToken: env.VITE_SENTRY_AUTH_TOKEN,
         org: "watchit",
         project: "watchit-app",
+        authToken: process.env.VITE_SENTRY_AUTH_TOKEN,
       }),
       nodePolyfills({
         // To add only specific polyfills, add them here. If no option is passed, adds all polyfills
         include: ['process', "module", "buffer"],
         globals: { global: true, process: true, Buffer: true },
+      }),
+      codecovVitePlugin({
+        bundleName: "watchit-app",
+        enableBundleAnalysis: process.env.VITE_CODECOV_TOKEN !== undefined,
+        uploadToken: process.env.VITE_CODECOV_TOKEN,
       }),
     ],
     resolve: {
@@ -40,6 +47,15 @@ export default defineConfig(({ mode }) => {
     define: {
       'process.env': env, // Make sure to define process.env for compatibility
     },
+
+    test: {
+      global: true,
+      environment: 'jsdom',
+      setupFiles: ['./setupTest.ts'],
+      coverage: {
+        provider: 'v8'
+      }
+    },
   };
-  
+
 });